lifestyle-web/server/NUT_model/requirements.txt
torch
torchvision
Pillow
opencv-python-headless
numpy
matplotlib
requests
